+++ js/entries/share-entry.js
@@ -1,4 +1,5 @@
 // Entry point for share.html
 import '../utils.js';
 import '../i18n.js';
+import '../activityCategories.js';
 import '../share.js';